Add files via upload
The deployment of the music recommendation system was carried out using PyCharm. The system was deployed with the help of the Streamlit module, which is designed for creating web applications quickly and intuitively. By utilizing Streamlit, the machine learning model and recommendation engine were integrated into an interactive web interface. This allowed users to input their preferences and receive real-time music recommendations through a user-friendly platform. Streamlit's simplicity and powerful capabilities facilitated the seamless translation of the backend recommendation logic into a frontend application, making the recommendation system accessible and easy to use.